+++ R/canvas_squares.R
@@ -3,14 +3,13 @@
 #' @description This function paints random squares and rectangles. It works by repeatedly cutting into the canvas at random locations and coloring the area that these cuts create.
 #'
 #' @usage canvas_squares(colors, background = "#000000", cuts = 50, ratio = 1.618,
-#'                width = 200, height = 200, noise = FALSE)
+#'                resolution = 200, noise = FALSE)
 #'
 #' @param colors      a string or character vector specifying the color(s) used for the artwork.
 #' @param background  a character specifying the color used for the borders of the squares.
 #' @param cuts        a positive integer specifying the number of cuts to make.
 #' @param ratio       a value specifying the \code{1:1} ratio for each cut.
-#' @param width       a positive integer specifying the width of the artwork in pixels.
-#' @param height      a positive integer specifying the height of the artwork in pixels.
+#' @param resolution  resolution of the artwork in pixels per row/column. Increasing the resolution increases the quality of the artwork but also increases the computation time exponentially.
 #' @param noise       logical. Whether to add k-nn noise to the artwork. Note that adding noise increases computation time significantly in large dimensions.
 #'
 #' @return A \code{ggplot} object containing the artwork.
@@ -32,18 +31,18 @@
 #' @export
 
 canvas_squares <- function(colors, background = "#000000", cuts = 50, ratio = 1.618,
-                           width = 200, height = 200, noise = FALSE) {
-  .checkUserInput(width = width, height = height, background = background)
+                           resolution = 200, noise = FALSE) {
+  .checkUserInput(resolution = resolution, background = background)
   if (cuts <= 1) {
     stop("Cuts must be higher than 1.")
   }
   palette <- c(background, colors)
   neighbors <- expand.grid(-1:1, -1:1)
   colnames(neighbors) <- c("x", "y")
-  canvas <- matrix(0, nrow = height, ncol = width)
+  canvas <- matrix(0, nrow = resolution, ncol = resolution)
   full_canvas <- draw_squares(canvas, neighbors, length(colors), cuts, ratio)
   if (noise) {
-    full_canvas <- full_canvas - .noise(dims = c(height, width))
+    full_canvas <- full_canvas - .noise(dims = c(resolution, resolution))
   }
   full_canvas <- .unraster(full_canvas, names = c("x", "y", "z"))
   artwork <- ggplot2::ggplot(data = full_canvas, ggplot2::aes(x = x, y = y, fill = z)) +